+++ efb/src/geojson/mod.rs
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-// Copyright 2024 Joe Pearson
+// Copyright 2025 Joe Pearson
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,11 +13,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Geometry.
-mod bbox;
-mod coordinate;
-mod polygon;
-
-pub(crate) use bbox::*;
-pub use coordinate::*;
-pub use polygon::*;
+mod geom;
+mod route;